+++ src/main/java/de/maxhenkel/gravestone/items/DeathInfoItem.java
@@ -34,7 +34,7 @@ public class DeathInfoItem extends Item {
     public ActionResult<ItemStack> onItemRightClick(World worldIn, PlayerEntity playerIn, Hand hand) {
         DeathInfo info = DeathInfo.getDeathInfoFromPlayerHand(playerIn);
 
-        if (playerIn.isCrouching() && playerIn.abilities.isCreativeMode) {
+        if (playerIn.func_225608_bj_() && playerIn.abilities.isCreativeMode) {
             if (playerIn instanceof ServerPlayerEntity) {
                 NetworkHooks.openGui((ServerPlayerEntity) playerIn, new INamedContainerProvider() {