+++ changelog.md
@@ -1,2 +1,1 @@
-- Updated to 1.15.1
-- Added configurable color for the gravestone
+- Fixed crouching issue with death info